+++ dist_node/cont.js
@@ -5,17 +5,17 @@
 "use strict";
 var Identity = require("./identity"),
     ContT = require("./trans/cont"),
-    runContT = ContT["runContT"],
-    Cont;
+    Cont, runContT = ContT["runContT"];
 (Cont = ContT(Identity));
 var x = (function(m, k) {
     var y;
-    return runContT(m, ((y = Identity.of), (function(x0) {
-        return y(k(x0));
+    return runContT(m, ((y = Identity.of), (function(z) {
+        return y(k(z));
     })));
 }),
     y = Identity.runIdentity;
 (Cont.runCont = (function() {
-    return y(x.apply(null, arguments));
+    var args = arguments;
+    return y(x.apply(null, args));
 }));
 (module.exports = Cont);