+++ dist_node/trans/list.js
@@ -4,15 +4,14 @@
 */
 "use strict";
 var __o = require("../base"),
-    liftM = __o["liftM"],
-    liftM2 = __o["liftM2"],
     __o0 = require("../structure"),
-    Functor = __o0["Functor"],
+    StateMonad = require("../spec/state"),
+    ListT, liftM = __o["liftM"],
+    liftM2 = __o["liftM2"],
     Monoid = __o0["Monoid"],
     Monad = __o0["Monad"],
     Transformer = __o0["Transformer"],
-    StateMonad = require("../spec/state"),
-    ListT, map = (function(f, a) {
+    map = (function(f, a) {
         return Array.prototype.map.call(a, f);
     }),
     concat = Function.prototype.call.bind(Array.prototype.concat),
@@ -32,7 +31,8 @@ var __o = require("../base"),
             return Array.prototype.reduceRight.call(a, f, z);
         })),
         mapM = (function() {
-            return sequence(map.apply(null, arguments));
+            var args = arguments;
+            return sequence(map.apply(null, args));
         });
     Monoid(Instance, new(Instance)(m.of([])), (function(b) {
         var a = this;
@@ -42,9 +42,9 @@ var __o = require("../base"),
         return new(Instance)(m.of([x]));
     }), (function(f0) {
         var c = this;
-        return new(Instance)(flattenM(c.run.chain(mapM.bind(null, (function(x) {
-            var x0 = f0(x);
-            return x0.run;
+        return new(Instance)(flattenM(c.run.chain(mapM.bind(null, (function(z0) {
+            var x = f0(z0);
+            return x.run;
         })))));
     }));
     Transformer(Instance, m, (function(t) {