+++ dist_node/trans/state.js
@@ -37,7 +37,7 @@ var x = StateT.runStateT,
 }));
 var x0 = StateT.runStateT,
     y0 = map.bind(null, (function(x1) {
-        return x1.x;
+        return x1.s;
     }));
 (StateT.execStateT = (function() {
     var args = arguments;